massive refactor of camera code camera now handles zoom properly sprites and polygons do not use camera to render themselves, they leave it to the projection matrix configurable camera_query_expansion parameter fixed gl_FragCoord layouts so that we dont have to revert y

+++ src/augs/filesystem/file.h
@@ -93,7 +93,7 @@ namespace augs {
 
 		std::vector<std::byte> output;
 
-		output.resize(size);
+		output.resize(static_cast<std::size_t>(size));
 		file.read(reinterpret_cast<byte_type_for_t<std::ifstream>*>(output.data()), size);
 		return output;
 	}
